[FIX] web: set the value before showing datetimepicker
Before this commit, the datetimepicker was shown without the correct input value
selected because the 'date' was set after the datetimepicker display.

This commit sets the 'viewDate' before showing the datetimepicker.
We use the 'viewDate' instead of 'date' because the lib sets the 'date' from
the input value. So the '_onShow' method was unecessary.